improve file organisation
