+++ src/View/Admin/_activityAdmin.html.twig
@@ -14,7 +14,7 @@
             {%  for activity in activities %}
                 <tr>
                     <td>{{ activity.name }}</td>
-                    <td><a class="mini activity-button btn btn-primary btn-lg" href="/./Admin/showActivity/{{ activity.id }}">Modifier</a>
+                    <td><a class="mini activity-button btn btn-primary btn-lg" href="/./adminActivity/showActivity/{{ activity.id }}">Modifier</a>
                     </td>
                     <td>
                         <button>
